+++ full-moon/src/ast/parsers.rs
@@ -1,5 +1,5 @@
 use super::{
-    parser_util::{InternalAstError, Parser},
+    parser_util::{InternalAstError, ParseIntoBox, Parser},
     span::ContainedSpan,
     *,
 };
@@ -255,39 +255,63 @@ define_parser!(ParseParenExpression, Expression, |_, state| {
 });
 
 #[derive(Clone, Debug, PartialEq)]
-struct ParseValueExpression;
-define_parser!(ParseValueExpression, Expression, |_, state| {
-    let (state, value) = keep_going!(ParseValue.parse(state))?;
-    #[cfg(feature = "roblox")]
-    let (state, type_assertion) =
-        if let Ok((state, type_assertion)) = keep_going!(ParseTypeAssertion.parse(state)) {
-            (state, Some(type_assertion))
-        } else {
-            (state, None)
-        };
-
-    let value = Box::new(value);
-
-    Ok((
-        state,
-        Expression::Value {
-            value,
-            #[cfg(feature = "roblox")]
-            type_assertion,
-        },
-    ))
-});
-
-#[derive(Clone, Debug, PartialEq)]
 struct ParsePartExpression;
+
+// This looks really stupid because we save 66% of the stack allocation cost
+// by doing this.
 define_parser!(ParsePartExpression, Expression, |_, state| {
-    if let Ok((state, expression)) = keep_going!(ParseUnaryExpression.parse(state)) {
-        Ok((state, expression))
-    } else if let Ok((state, expression)) = keep_going!(ParseValueExpression.parse(state)) {
-        Ok((state, expression))
-    } else {
-        Err(InternalAstError::NoMatch)
+    #[allow(clippy::result_large_err)]
+    fn parse_unary_expression(
+        state: ParserState,
+    ) -> Result<(ParserState, Expression), InternalAstError> {
+        ParseUnaryExpression.parse(state)
     }
+
+    #[allow(clippy::result_large_err)]
+    fn parse_value(state: ParserState) -> Result<(ParserState, Expression), InternalAstError> {
+        ParseValue.parse(state)
+    }
+
+    #[allow(clippy::result_large_err)]
+    fn parse_paren_expression(
+        state: ParserState,
+    ) -> Result<(ParserState, Expression), InternalAstError> {
+        ParseParenExpression.parse(state)
+    }
+
+    let (state, expression) = keep_going!(parse_unary_expression(state)).or_else(|_| {
+        keep_going!(parse_value(state)).or_else(|_| keep_going!(parse_paren_expression(state)))
+    })?;
+
+    #[cfg(feature = "roblox")]
+    #[allow(clippy::result_large_err)]
+    fn maybe_type_assertion(
+        state: ParserState,
+        expression: Expression,
+    ) -> Result<(ParserState, Expression), InternalAstError> {
+        if let Ok((state, type_assertion)) = keep_going!(ParseTypeAssertion.parse(state)) {
+            return Ok((
+                state,
+                Expression::TypeAssertion {
+                    expression: Box::new(expression),
+                    type_assertion,
+                },
+            ));
+        }
+
+        Ok((state, expression))
+    }
+
+    #[cfg(not(feature = "roblox"))]
+    #[allow(clippy::result_large_err)]
+    fn maybe_type_assertion(
+        state: ParserState,
+        expression: Expression,
+    ) -> Result<(ParserState, Expression), InternalAstError> {
+        Ok((state, expression))
+    }
+
+    maybe_type_assertion(state, expression)
 });
 
 #[derive(Clone, Debug, PartialEq)]
@@ -352,22 +376,21 @@ define_roblox_parser!(
 
 #[derive(Clone, Debug, PartialEq)]
 struct ParseValue;
-define_parser!(ParseValue, Value, |_, state| parse_first_of!(state, {
-    ParseSymbol(Symbol::Nil) => Value::Symbol,
-    ParseSymbol(Symbol::False) => Value::Symbol,
-    ParseSymbol(Symbol::True) => Value::Symbol,
-    ParseNumber => Value::Number,
-    ParseStringLiteral => Value::String,
-    ParseSymbol(Symbol::Ellipse) => Value::Symbol,
-    ParseFunction => Value::Function,
-    ParseTableConstructor => Value::TableConstructor,
-    ParseFunctionCall => Value::FunctionCall,
-    ParseVar => Value::Var,
-    ParseParenExpression => Value::ParenthesesExpression,
+define_parser!(ParseValue, Expression, |_, state| parse_first_of!(state, {
+    ParseSymbol(Symbol::Nil) => Expression::Symbol,
+    ParseSymbol(Symbol::False) => Expression::Symbol,
+    ParseSymbol(Symbol::True) => Expression::Symbol,
+    ParseNumber => Expression::Number,
+    ParseStringLiteral => Expression::String,
+    ParseSymbol(Symbol::Ellipse) => Expression::Symbol,
+    ParseFunction => Expression::Function,
+    ParseTableConstructor => Expression::TableConstructor,
+    ParseFunctionCall => Expression::FunctionCall,
+    ParseVar => Expression::Var,
     @#[cfg(feature = "roblox")]
-    ParseIfExpression => Value::IfExpression,
+    ParseIfExpression => Expression::IfExpression,
     @#[cfg(feature = "roblox")]
-    ParseInterpolatedString => Value::InterpolatedString,
+    ParseInterpolatedString => Expression::InterpolatedString,
 }));
 
 #[derive(Clone, Debug, Default, PartialEq)]
@@ -399,7 +422,7 @@ define_parser!(ParseStmt, Stmt, |_, state| parse_first_of!(state, {
 #[derive(Clone, Debug, PartialEq)]
 struct ParsePrefix;
 define_parser!(ParsePrefix, Prefix, |_, state| parse_first_of!(state, {
-    ParseParenExpression => Prefix::Expression,
+    ParseIntoBox(ParseParenExpression) => Prefix::Expression,
     ParseIdentifier => Prefix::Name,
 }));
 
@@ -871,12 +894,12 @@ define_parser!(ParseSuffix, Suffix, |_, state| parse_first_of!(state, {
 
 #[derive(Clone, Debug, PartialEq)]
 struct ParseVarExpression;
-define_parser!(ParseVarExpression, VarExpression, |_, state| {
+define_parser!(ParseVarExpression, Box<VarExpression>, |_, state| {
     let (state, prefix) = ParsePrefix.parse(state)?;
     let (state, suffixes) = ZeroOrMore(ParseSuffix).parse(state)?;
 
     if let Some(Suffix::Index(_)) = suffixes.last() {
-        Ok((state, VarExpression { prefix, suffixes }))
+        Ok((state, Box::new(VarExpression { prefix, suffixes })))
     } else {
         Err(InternalAstError::NoMatch)
     }
@@ -1430,12 +1453,12 @@ cfg_if::cfg_if! {
                     state,
                     IfExpression {
                         if_token,
-                        condition,
+                        condition: Box::new(condition),
                         then_token,
-                        if_expression,
+                        if_expression: Box::new(if_expression),
                         else_if_expressions: if else_if_expressions.is_empty() { None } else { Some(else_if_expressions) },
                         else_token,
-                        else_expression,
+                        else_expression: Box::new(else_expression),
                     },
                 ))
             }